Create AdminOpeningHourController. Move everything except widget
