+++ app/Resources/views/admin/openinghour/widget_generator.html.twig
@@ -44,12 +44,12 @@
     <h4>Résultat</h4>
     <div class="row">
         <div class="col m4">
-            <iframe src="{{ path('admin_openinghour_widget') }}?{{ query_string }}" frameborder="0" style="border: 2px solid green;"></iframe>
+            <iframe src="{{ path('openinghour_widget') }}?{{ query_string }}" frameborder="0" style="border: 2px solid green;"></iframe>
         </div>
         <div class="col m8">
             texte à copier-coller 👇
             <div class="card-panel" style="overflow:scroll;">
-                <pre>&lt;iframe src="{{ absolute_url(path('admin_openinghour_widget')) }}?{{ query_string }}" frameborder="0"&gt;&lt;/iframe&gt;</pre>
+                <pre>&lt;iframe src="{{ absolute_url(path('openinghour_widget')) }}?{{ query_string }}" frameborder="0"&gt;&lt;/iframe&gt;</pre>
             </div>
         </div>
     </div>
